Fix EZP-22840: Editor user missing access to see frontend
Actually fixing issues introduced in EZP-22550 "Added breadcrumb to
navigate in Demobundle" by making sure location 1 (which editor does
not have access to) is not loaded as part of bread crumbs.

Could be argued that Editor should inherit Anonymous role to get
access, but that is currently not the case, and besides this saves
a uneeded load call either way.